ref: Removing unnecessary files
