+++ src/router.js
@@ -1,77 +1,52 @@
 import { createRouter, createWebHistory } from "vue-router";
 
-import Login from "./views/Login.vue";
-import Home from "./views/Home.vue";
-// import TutorialsList from "./views/TutorialsList.vue";
-// import EditTutorial from "./views/EditTutorial.vue";
-// import AddTutorial from "./views/AddTutorial.vue";
-// import ViewTutorial from "./views/ViewTutorial.vue";
-// import AddLesson from "./views/AddLesson.vue";
-// import EditLesson from "./views/EditLesson.vue";
-import AllAssets from "./views/AllAssets.vue";
-import AssetData from "./views/AssetData.vue";
-import AssetTypeManagement from "./views/AssetTypeManagement.vue";
-import AssetCatManagement from "./views/ManageAssetCategories.vue";
-import AssetTemplateManagement from "./views/AssetTemplateManagement.vue";
-import AssetFieldManagement from "./views/AssetFieldManagement.vue";
-import BuildingManagement from "./views/BuildingManagement.vue";
-import GroupManagement from "./views/GroupManagement.vue";
-import PersonManagement from "./views/PersonManagement.vue";
-import ViewPerson from './views/EditPerson.vue';
-import RoomManagement from "./views/RoomManagement.vue";
-import TemplateDataManagement from "./views/TemplateDataManagement.vue";
-import VendorManagement from "./views/VendorManagement.vue";
-import Alerts from "./views/AlertManagement.vue";
 import AlertTypeManagement from "./views/AlertTypeManagement.vue";
-import AddEditAssetType from "./views/EditAssetType.vue";
-import FieldList from "./views/fieldLists.vue";
-import FieldListOptions from "./views/fieldListOptions.vue";
-import Logs from "./views/logs.vue";
-import Notifications from "./views/notificationsManagement.vue";
-import Permissions from "./views/permissionManagement.vue";
-import Users from "./views/userManagement.vue";
-import GenerateReports from "./views/GenerateReports.vue";
+import AllAssets from "./views/AllAssets.vue";
+import AssetTypeManagement from "./views/AssetTypeManagement.vue";
+import AssetCatManagement from "./views/AssetCatManagement.vue";
+import AssetTemplateManagement from "./views/AssetTemplateManagement.vue";
+import BuildingManagement from "./views/BuildingManagement.vue";
 import EditAssetType from "./views/EditAssetType.vue";
-
-// import TutorialsList from "./views/TutorialsList.vue";
-// import EditTutorial from "./views/EditTutorial.vue";
-// import AddTutorial from "./views/AddTutorial.vue";
-// import ViewTutorial from "./views/ViewTutorial.vue";
-// import AddLesson from "./views/AddLesson.vue";
-// import EditLesson from "./views/EditLesson.vue";
-// import AllAssets from "./views/allAssets.vue";
+import FieldListManagement from "./views/FieldListManagement.vue";
+import GenerateReports from "./views/GenerateReports.vue";
+import GroupManagement from "./views/GroupManagement.vue";
+import Home from "./views/Home.vue";
+import Login from "./views/Login.vue";
+import PersonManagement from "./views/PersonManagement.vue";
+import VendorManagement from "./views/VendorManagement.vue";
+import ViewPerson from './views/EditPerson.vue';
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
   routes: [
     {
       path: "/",
-      alias: "/login",
+      alias: "/home",
+      name: "home",
+      component: Home,
+    },
+    {
+      path: "/login",
       name: "login",
       component: Login,
     },
     {
-      path: "/home",
-      name: "home",
-      component: Home,
+      path: "/alert-types",
+      name: "alert-types",
+      component: AlertTypeManagement,
     },
     {
       path: "/assets",
       name: "assets",
       component: AllAssets,
     },
-    {
-      path: "/asset-data",
-      name: "asset-data",
-      component: AssetData,
-    },
-    {
-      path: "/asset-type",
-      name: "asset-type",
+    {//---------------------------------------------------------------------------------------------------
+      path: "/asset-types",
+      name: "asset-types",
       component: AssetTypeManagement,
     },
     {
-      path: "/asset-type/:id",
+      path: "/asset-types/:id",
       name: "asset-type-edit",
       component: EditAssetType,
       props: true,
@@ -87,14 +62,14 @@ const router = createRouter({
       component: AssetTemplateManagement,
     },
     {
-      path: "/asset-fields",
-      name: "asset-fields",
-      component: AssetFieldManagement,
-    },
-    {
       path: "/buildings",
       name: "buildings",
       component: BuildingManagement,
+    },
+    {
+      path: "/field-list",
+      name: "field-list",
+      component: FieldListManagement,
     },
     {
       path: "/groups",
@@ -107,7 +82,7 @@ const router = createRouter({
       component: PersonManagement,
     },
     {
-      path: "/person/:id", 
+      path: "/people/:id", 
       name: "person",
       component: ViewPerson,
       props: true,
@@ -118,94 +93,10 @@ const router = createRouter({
       component: GenerateReports,
     },
     {
-      path: "/rooms",
-      name: "rooms",
-      component: RoomManagement,
-    },
-    {
-      path: "/template-data",
-      name: "template-data",
-      component: TemplateDataManagement,
-    },
-    {
       path: "/vendors",
       name: "vendors",
       component: VendorManagement,
     },
-    {
-      path: "/alerts",
-      name: "alerts",
-      component: Alerts,
-    },
-    {
-      path: "/alert-types",
-      name: "alert-types",
-      component: AlertTypeManagement,
-    },
-    {
-      path: "/field-list",
-      name: "field-list",
-      component: FieldList,
-    },
-    {
-      path: "/field-list-options",
-      name: "field-list-options",
-      component: FieldListOptions,
-    },
-    {
-      path: "/logs",
-      name: "logs",
-      component: Logs,
-    },
-    {
-      path: "/notifications",
-      name: "notifications",
-      component: Notifications,
-    },
-    {
-      path: "/permissions",
-      name: "permissions",
-      component: Permissions,
-    },
-    {
-      path: "/users",
-      name: "users",
-      component: Users,
-    },
-    //{
-    //   path: "/tutorials",
-    //   name: "tutorials",
-    //   component: TutorialsList,
-    // },
-    // {
-    //   path: "/edit/:id",
-    //   name: "edit",
-    //   component: EditTutorial,
-    //   props: true,
-    // },
-    // {
-    //   path: "/add",
-    //   name: "add",
-    //   component: AddTutorial,
-    // },
-    // {
-    //   path: "/view/:id",
-    //   name: "view",
-    //   component: ViewTutorial,
-    //   props: true,
-    // },
-    // {
-    //   path: "/addLesson/:tutorialId",
-    //   name: "addLesson",
-    //   component: AddLesson,
-    //   props: true,
-    // },
-    // {
-    //   path: "/editLesson/:tutorialId/:lessonId",
-    //   name: "editLesson",
-    //   component: EditLesson,
-    //   props: true,
-    // },
   ],
 });
