--- src/views/AssetData.vue
@@ -1,8 +1,0 @@
-<script setup>
-</script>
-
-<template>
-</template>
-
-<style>
-</style>
