--- src/views/TutorialsList.vue
@@ -1,86 +1,0 @@
-<script setup>
-import TutorialServices from "../services/tutorialServices";
-import Utils from "../config/utils.js";
-import { ref } from "vue";
-import { useRouter } from "vue-router";
-
-const router = useRouter();
-const tutorials = ref([]);
-const user = Utils.getStore("user");
-const message = ref("Search, Edit or Delete Tutorials");
-
-const editTutorial = (tutorial) => {
-  router.push({ name: "edit", params: { id: tutorial.id } });
-};
-
-const viewTutorial = (tutorial) => {
-  router.push({ name: "view", params: { id: tutorial.id } });
-};
-
-const deleteTutorial = (tutorial) => {
-  TutorialServices.delete(tutorial.id)
-    .then(() => {
-      retrieveTutorials();
-    })
-    .catch((e) => {
-      message.value = e.response.data.message;
-    });
-};
-
-const retrieveTutorials = () => {
-  TutorialServices.getAllForUser(user.userId)
-    .then((response) => {
-      tutorials.value = response.data;
-    })
-    .catch((e) => {
-      message.value = e.response.data.message;
-    });
-};
-
-retrieveTutorials();
-</script>
-
-<template>
-  <div>
-    <v-container>
-      <v-toolbar>
-        <v-toolbar-title
-          >Hello, {{ user.fName }} {{ user.lName }}!</v-toolbar-title
-        >
-      </v-toolbar>
-      <br /><br />
-      <v-card>
-        <v-card-title> Tutorials </v-card-title>
-        <v-card-text>
-          <b>{{ message }}</b>
-        </v-card-text>
-        <v-table>
-          <thead>
-            <tr>
-              <th class="text-left">Title</th>
-              <th class="text-left">Description</th>
-              <th class="text-left">Actions</th>
-            </tr>
-          </thead>
-          <tbody>
-            <tr v-for="(item, index) in tutorials" :key="item.title">
-              <td>{{ item.title }}</td>
-              <td>{{ item.description }}</td>
-              <td>
-                <v-icon small class="mx-4" @click="editTutorial(item)">
-                  mdi-pencil
-                </v-icon>
-                <v-icon small class="mx-4" @click="viewTutorial(item)">
-                  mdi-format-list-bulleted-type
-                </v-icon>
-                <v-icon small class="mx-4" @click="deleteTutorial(item)">
-                  mdi-trash-can
-                </v-icon>
-              </td>
-            </tr>
-          </tbody>
-        </v-table>
-      </v-card>
-    </v-container>
-  </div>
-</template>
