--- src/views/fieldListOptions.vue
@@ -1,9 +1,0 @@
-<script setup>
-</script>
-
-<template>
-
-</template>
-
-<style>
-</style>
